Refactor PolarizationCorrection tests
Created an abstract class for the more commonly used methods.

RE #37225
